v2.0/requirements.txt
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0
paramiko>=3.0.0
pypdf>=6.0.0
easyocr>=1.7.0
playwright>=1.40.0
selenium>=4.0.0
webdriver-manager>=4.0.0
